ui/requirements.txt
streamlit>=1.30.0
httpx>=0.26.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
sseclient-py>=1.8.0
